whatsapp_bot/requirements.txt
# Aremu WhatsApp Bot Dependencies

# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# AI & OpenAI
openai

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Scheduling
schedule==1.2.0

# Database
psycopg2-binary==2.9.7

# Vector Database & Embeddings
pgvector==0.2.4
numpy==1.24.3
scikit-learn==1.3.0

# Scraper Dependencies
python-jobspy>=1.1.0
pandas>=1.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
schedule>=1.2.0

# JSON handling (built-in, but explicit for clarity)
# json - built-in

# Logging (built-in)
# logging - built-in

# Date/Time handling (built-in)
# datetime - built-in

# Security & Hashing (built-in)
# hashlib - built-in
